@@ -1,24 +1,26 @@
 # -*- encoding: utf-8 -*-
-lib = File.expand_path('../lib', __FILE__)
-$LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
+# frozen_string_literal: true
+
+$:.push File.expand_path('../lib', __FILE__)
 require 'omniauth-polaris/version'
+
 Gem::Specification.new do |gem|
-  gem.name = "omniauth-polaris"
-  gem.version = OmniAuth::Polaris::VERSION
+  gem.name = 'omniauth-polaris'
+  gem.version = OmniAuth::Polaris::VERSION.dup
   gem.authors = ["Steven Anderson", "Ben Barber"]
   gem.email = ["sanderson@bpl.org", "bbarber@bpl.org"]
   gem.description = %q{A Polaris API strategy for OmniAuth.}
   gem.summary = %q{A Polaris API strategy for OmniAuth.}
   gem.homepage = "https://github.com/boston-library/omniauth-polaris"
 
-  gem.required_ruby_version = '~> 2.4'
+  gem.required_ruby_version = '>= 2.6'
 
+  gem.add_dependency 'http', '~> 5.1'
+  gem.add_dependency 'omniauth', '>= 2.0.0'
 
-  gem.add_dependency 'rest-client', '~> 2.0.2'
-
-  gem.add_development_dependency 'omniauth', '~> 1.8.1'
   gem.add_development_dependency 'bundler', '>= 1.3.0'
   gem.add_development_dependency 'awesome_print'
+  gem.add_development_dependency 'pry'
   gem.add_development_dependency 'rspec', '~> 3.8'
   gem.add_development_dependency 'libnotify', '~> 0.9.3'
 
