@@ -1,3 +1,4 @@
-require 'omniauth-polaris/version'
-require 'omniauth-polaris/adaptor'
-require 'omniauth/strategies/polaris'
+# frozen_string_literal: true
+
+require_relative './omniauth/polaris/version'
+require_relative './omniauth/polaris'
